fix and organize csv file functions
