+++ app/(tabs)/settings.tsx
@@ -8,7 +8,6 @@ import { ScrollView } from "react-native-gesture-handler";
 import { IWord } from "@/interfaces/languageObjectInterface";
 import { SafeAreaView } from "react-native-safe-area-context";
 import styles from "@/assets/styles/styleSheet";
-import { saveDataToCSV } from "@/file-management/movingFiles";
 import ManageFileModal from "@/components/ManageFileModal";
 import { asyncStorageGetAllKeys } from "@/utilities/utility-async-storage";
